+++ docs/source/CHANGELOG.rst
@@ -9,13 +9,13 @@ Primary (X.-.-) version numbers are used to denote backwards
 incompatibilities between versions, while minor (-.X.-) numbers
 primarily indicate new features and documentation.
 
-5.0.0-per3 (XXXX-XX-XX)
+5.0.0-per3 (2025-04-25)
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
 * Remove 'GStepBStdevModel', which is useless.  Use GstepBModel. 
 
 
-5.0.0-pre2 (2024-05-24)
+5.0.0-pre2 (2025-03-24)
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
 * Fix problem in kz-conversion.py